+++ BraintreeModels/src/test/java/com/braintreepayments/api/PaymentMethodNonceUnitTest.java
@@ -20,7 +20,6 @@ public class PaymentMethodNonceUnitTest {
         assertTrue(nonce.isDefault());
         assertEquals(PaymentMethodType.CARD, nonce.getType());
         assertEquals("123456-12345-12345-a-adfa", nonce.getString());
-        assertEquals("ending in ••11", nonce.getDescription());
     }
 
     @Test
@@ -30,7 +29,6 @@ public class PaymentMethodNonceUnitTest {
         assertFalse(nonce.isDefault());
         assertEquals(PaymentMethodType.PAYPAL, nonce.getType());
         assertEquals("aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee", nonce.getString());
-        assertEquals("with email paypalaccount@example.com", nonce.getDescription());
     }
 
     @Test
@@ -40,7 +38,6 @@ public class PaymentMethodNonceUnitTest {
         assertFalse(nonce.isDefault());
         assertEquals(PaymentMethodType.VENMO, nonce.getType());
         assertEquals("fake-venmo-nonce", nonce.getString());
-        assertEquals("VenmoAccount", nonce.getDescription());
     }
 
     @Test
@@ -50,7 +47,6 @@ public class PaymentMethodNonceUnitTest {
         assertFalse(nonce.isDefault());
         assertEquals(PaymentMethodType.VISA_CHECKOUT, nonce.getType());
         assertEquals("123456-12345-12345-a-adfa", nonce.getString());
-        assertEquals("ending in ••11", nonce.getDescription());
     }
 
     @Test
@@ -60,7 +56,6 @@ public class PaymentMethodNonceUnitTest {
         assertFalse(nonce.isDefault());
         assertEquals(PaymentMethodType.PAYPAL, nonce.getType());
         assertEquals("e11c9c39-d6a4-0305-791d-bfe680ef2d5d", nonce.getString());
-        assertEquals("PayPal", nonce.getDescription());
     }
 
     @Test
@@ -70,6 +65,5 @@ public class PaymentMethodNonceUnitTest {
         assertFalse(nonce.isDefault());
         assertEquals(PaymentMethodType.GOOGLE_PAY, nonce.getType());
         assertEquals("fake-google-pay-nonce", nonce.getString());
-        assertEquals("Google Pay", nonce.getDescription());
     }
 }