+++ PayPal/src/test/java/com/braintreepayments/api/PayPalAccountNonceUnitTest.java
@@ -24,7 +24,6 @@ public class PayPalAccountNonceUnitTest {
         assertNotNull(payPalAccountNonce);
         assertEquals(PaymentMethodType.PAYPAL, payPalAccountNonce.getType());
         assertEquals("fake-authenticate-url", payPalAccountNonce.getAuthenticateUrl());
-        assertEquals("with email paypalaccount@example.com", payPalAccountNonce.getDescription());
         assertEquals("aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee", payPalAccountNonce.getString());
         assertEquals("paypalaccount@example.com", payPalAccountNonce.getEmail());
         assertEquals("123 Fake St.", payPalAccountNonce.getBillingAddress().getStreetAddress());
@@ -53,7 +52,6 @@ public class PayPalAccountNonceUnitTest {
 
         assertNotNull(payPalAccountNonce);
         assertEquals(PaymentMethodType.PAYPAL, payPalAccountNonce.getType());
-        assertEquals("with email paypalaccount@example.com", payPalAccountNonce.getDescription());
         assertEquals("aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee", payPalAccountNonce.getString());
         assertEquals("paypalaccount@example.com", payPalAccountNonce.getEmail());
         assertEquals("123 Fake St.", payPalAccountNonce.getBillingAddress().getStreetAddress());
@@ -79,7 +77,6 @@ public class PayPalAccountNonceUnitTest {
     public void getDescription_usesGetEmailIfDescriptionIsPayPalAndEmailIsNotEmpty() throws JSONException {
         PayPalAccountNonce payPalAccountNonce =
             PayPalAccountNonce.fromJSON(new JSONObject(Fixtures.PAYMENT_METHODS_PAYPAL_ACCOUNT_RESPONSE_WITH_DEFAULT_DESCRIPTION));
-        assertEquals("paypalaccount@example.com", payPalAccountNonce.getDescription());
     }
 
     @Test
@@ -92,7 +89,6 @@ public class PayPalAccountNonceUnitTest {
         PayPalAccountNonce parceled = PayPalAccountNonce.CREATOR.createFromParcel(parcel);
 
         assertEquals(PaymentMethodType.PAYPAL, payPalAccountNonce.getType());
-        assertEquals("with email paypalaccount@example.com", parceled.getDescription());
         assertEquals("aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee", parceled.getString());
         assertEquals("paypalaccount@example.com", parceled.getEmail());
         assertEquals("fake-authenticate-url", parceled.getAuthenticateUrl());
@@ -131,7 +127,6 @@ public class PayPalAccountNonceUnitTest {
         assertNull(parceled.getCreditFinancing());
 
         assertEquals(PaymentMethodType.PAYPAL, payPalAccountNonce.getType());
-        assertEquals("with email paypalaccount@example.com", parceled.getDescription());
         assertEquals("aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee", parceled.getString());
         assertEquals("paypalaccount@example.com", parceled.getEmail());
         assertEquals("123 Fake St.", parceled.getBillingAddress().getStreetAddress());